+++ src/NTangle/Events/EventSourceFormatter.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/NTangle
 
 using CoreEx.Entities;
+using CoreEx.Events;
 using System;
 
 namespace NTangle.Events
@@ -13,11 +14,12 @@ namespace NTangle.Events
         /// <summary>
         /// Formats the <paramref name="source"/> based on the specified <paramref name="format"/>.
         /// </summary>
+        /// <param name="formatter">The <see cref="EventActionFormatter"/>.</param>
         /// <param name="source">The source <see cref="Uri"/>.</param>
         /// <param name="value">The entity value.</param>
         /// <param name="format">The <see cref="EventSourceFormat"/>.</param>
         /// <returns>The formatted action.</returns>
-        public static Uri? Format<T>(Uri? source, T value, EventSourceFormat? format = EventSourceFormat.NameAndTableKey) where T : IEntityKey
+        public static Uri? Format<T>(EventDataFormatter formatter, Uri? source, T value, EventSourceFormat? format = EventSourceFormat.NameAndTableKey) where T : IEntityKey
         {
             if (source == null)
                 return null;
@@ -25,8 +27,8 @@ namespace NTangle.Events
             return format switch
             {
                 EventSourceFormat.NameOnly => source,
-                EventSourceFormat.NameAndKey => CreateUri(source, value.EntityKey.ToString()),
-                EventSourceFormat.NameAndTableKey => CreateUri(source, value is IGlobalIdentifier gi ? gi.TableKey.ToString() : value.EntityKey.ToString()),
+                EventSourceFormat.NameAndKey => CreateUri(source, value.EntityKey.ToString(formatter.KeySeparatorCharacter)),
+                EventSourceFormat.NameAndTableKey => CreateUri(source, value is IGlobalIdentifier gi ? gi.TableKey.ToString(formatter.KeySeparatorCharacter) : value.EntityKey.ToString(formatter.KeySeparatorCharacter)),
                 _ => null
             };
         }